Allow user to chose gem5 config script e.g., se.py
Other changes:
Spinner added, other minor improvements

+++ arch/arm/run_arm_se_benchmarks.sh
@@ -32,8 +32,33 @@
 #
 # Author: Éder F. Zulian
 
-source ../../common/defaults.in
-source ../../common/util.in
+DIR="$(cd "$(dirname "$0")" && pwd)"
+TOPDIR=$DIR/../..
+source $TOPDIR/common/defaults.in
+source $TOPDIR/common/util.in
+
+# Set to "yes" or "no" in order to enable a spinner or not.
+spinner="yes"
+
+# gem5 configuration script to be used. se.py, starter_se.py, etc.
+script="se.py"
+
+ncores="1"
+currtime=$(date "+%Y.%m.%d-%H.%M.%S")
+output_rootdir="se_output_$currtime"
+#tlm_options="--tlm-memory=transactor"
+
+if [ "$script" == "starter_se.py" ]; then
+	config_script="configs/example/arm/starter_se.py"
+	cpu_options="--cpu=hpi --num-cores=$ncores"
+	mem_options="--mem-channels=1"
+elif [ "$script" == "se.py" ]; then
+	config_script="configs/example/se.py"
+	cpu_options="--cpu-type=TimingSimpleCPU --num-cpu=$ncores"
+	mem_options="--mem-size=512MB --mem-channels=1 --caches --l2cache --mem-type=SimpleMemory"
+else
+	echo && echo "Please define options for $script" && echo && exit
+fi
 
 arch="ARM"
 mode="opt"
@@ -61,16 +86,11 @@ apps=(
 "test-suite/SingleSource/Benchmarks/McGill:queens"
 )
 
-config_script="configs/example/arm/starter_se.py"
-ncores="1"
-cpu_options="--cpu=hpi --num-cores=$ncores"
-currtime=$(date "+%Y.%m.%d-%H.%M.%S")
-output_rootdir="se_output_$currtime"
-mem_options="--mem-channels=1"
-#tlm_options="--tlm-memory=transactor"
-
-pulse &
-pupid=$!
+# start spinner
+if [ "$spinner" = "yes" ]; then
+	pulse &
+	pupid=$!
+fi
 declare -a pids
 for e in "${apps[@]}"; do
 	b=${e#*:}
@@ -83,8 +103,18 @@ for e in "${apps[@]}"; do
 	for ((c = 0; c < ncores; c++)); do
 		workload+="$wl "
 	done
-	$gem5_elf -d $output_dir $config_script $cpu_options $mem_options $tlm_options $workload > $logfile 2>&1 & pids+=($!)
+	if [ "$script" == "starter_se.py" ]; then
+		cmd="$gem5_elf -d $output_dir $config_script $cpu_options $mem_options $tlm_options $workload"
+	elif [ "$script" == "se.py" ]; then
+		cmd="$gem5_elf -d $output_dir $config_script -c $workload $mem_options $cpu_options $tlm_options"
+	else
+		echo && echo "Please define the command line for $script" && echo && exit
+	fi
+	$cmd > $logfile 2>&1 & pids+=($!)
 done
 wait "${pids[@]}"
 unset pids
-kill $pupid &>/dev/null
+# stop spinner
+if [ "$spinner" = "yes" ]; then
+	kill $pupid &>/dev/null
+fi
